+++ CHANGELOG.md
@@ -1,4 +1,8 @@
+## [0.0.12] - 09/03/2020.
+
+- Fixed to title case as AP-style based.
+
 ## [0.0.11] - 08/27/2020.
 
-* Fixed the remove duplicate white space function of an edge case 
-where sometimes it removes all new lines.
+- Fixed the remove duplicate white space function of an edge case
+  where sometimes it removes all new lines.